+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Revert smart rename affected files on cancel
 - Inconsistent spacing in configuration UI
 - Invalid revision when cancelling edit mode without changes
+- Set prompt library defaults on first language model added
 
 ## [2024.10.1] - 2024-09-05